try to compile passed

+++ src/unix/timer.cpp
@@ -20,7 +20,7 @@
 
 #include "uv.h"
 #include "internal.h"
-#include "heap-inl.h"
+#include "../heap-inl.h"
 
 #include <assert.h>
 #include <limits.h>
@@ -143,7 +143,7 @@ int uv__next_timeout(const uv_loop_t* loop) {
   if (diff > INT_MAX)
     diff = INT_MAX;
 
-  return diff;
+  return (int)diff;
 }
